+++ bld/browser/dlgprs/h/dialog.h
@@ -53,7 +53,7 @@ public:
                                         Dialog( const char * name, Rect rect );
                                         ~Dialog();
 
-            int                         addRect( const Rect & r );
+            size_t                      addRect( const Rect & r );
             void                        setCaption( const char * text );
             void                        setFont( int ptsize, const char * text );
             void                        setStyle( long style );
