+++ bld/gui/h/guix.h
@@ -92,9 +92,9 @@ extern bool GUIXCreateToolBarWithTips( gui_window *wnd, bool fixed, gui_ord heig
 
 extern bool GUIXCloseToolBar( gui_window *wnd );
 
-extern void GUIXDrawText( gui_window *wnd, const char *text, int length, gui_coord *pos,
+extern void GUIXDrawText( gui_window *wnd, const char *text, size_t length, gui_coord *pos,
                           gui_attr attr, gui_ord extentx, bool draw_extent );
-extern void GUIXDrawTextRGB( gui_window *wnd, char *text, int length, gui_coord *pos,
+extern void GUIXDrawTextRGB( gui_window *wnd, char *text, size_t length, gui_coord *pos,
                              gui_rgb fore, gui_rgb back, gui_ord extentx, bool draw_extent );
 extern void GUISetKeyState( void );
 extern gui_window *GUIXGetRootWindow( void );
