+++ bld/wclass/cpp/wname.cpp
@@ -40,8 +40,8 @@ WEXPORT WName::WName( WWindow* parent, const WRect& r, WWindow* client,
 /***************************/
 
     if( text != NULL ) {
-        int icount = strlen( text );
-        for( int i=0; i<icount; i++ ) {
+        size_t icount = strlen( text );
+        for( size_t i=0; i<icount; i++ ) {
             if( text[i] == '&' ) {
                 int key = text[i+1];
                 if( (key != '\0') && (key != '&') ) {
